Добавляет модульность в код проекта
