+++ js/config/settings.js
@@ -1,0 +1,17 @@
+const SETTINGS = {
+  SIMILAR_ADV_COUNT: 10, // Сколько объявлений генерировать
+  PRICE: {
+    MIN: 100000, // Минимальная стоимость жилья
+    MAX: 10000000 // Максимальная стоимость жилья
+  },
+  ROOMS: {
+    MIN: 1, // Минимальное количество комнат
+    MAX: 8 // Максимальное количество комнат
+  },
+  GUEST: {
+    MIN: 2, // Минимальное количество гостей для размещения
+    MAX: 14 // Максимальное количество гостей для размещения
+  },
+  DIGITS: 5 // Количество знаков после запятой
+};
+export {SETTINGS};
